bot/requirements.txt
aiodns==3.0.0
aiogram==2.13
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
brotlipy==0.7.0
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
greenlet==1.1.0
idna==3.2
importlib-metadata==4.5.0
multidict==5.1.0
Pillow==8.2.0
pycares==4.0.0
pycparser==2.20
pytz==2021.1
qrcode==6.1
six==1.16.0
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
ujson==4.0.2
yarl==1.6.3
zipp==3.4.1